requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-WTF==1.1.1
python-dotenv==1.0.0
firebase-admin==6.2.0
requests==2.31.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.97.0
dropbox==11.36.0
tenacity==8.2.3
openai==1.3.5
sentence-transformers==2.2.2
chromadb==0.4.14
PyPDF2==3.0.1
python-magic-bin==0.4.14; platform_system == "Windows"
python-magic==0.4.27; platform_system != "Windows"
python-docx==0.8.11
pandas==2.0.3
openpyxl==3.1.2